4.0 Worked on the Data Layer. Build the Network class, Service class, Request manager, network result handler.
Switched branches before this can a breaking functionality in the stable code we are having.

RequestService - Get the response from the server in the form of data

Request handler - Parse the data and stores it into the generic type for passing it to the service class.

FactsService - Interprets the generic type as FactsInfo and finally viewModel utilizes it. My next task is making Presentation Layer

Added reachability Library